Rename useTableClose to useWidgetClose
