+++ packages/jsapi-components/src/spectrum/utils/usePickerProps.ts
@@ -15,7 +15,7 @@ import { getItemKeyColumn, getItemLabelColumn } from './itemUtils';
 import { useItemRowDeserializer } from './useItemRowDeserializer';
 import { useGetItemIndexByValue } from '../../useGetItemIndexByValue';
 import useSearchableViewportData from '../../useSearchableViewportData';
-import useTableClose from '../../useTableClose';
+import useWidgetClose from '../../useWidgetClose';
 
 const log = Log.module('jsapi-components.usePickerProps');
 
@@ -78,7 +78,7 @@ export function usePickerProps<TProps>({
     [tableSource]
   );
 
-  useTableClose(tableCopy);
+  useWidgetClose(tableCopy);
 
   const keyColumn = useMemo(
     () =>
